# Auth/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Тестовое задание от MEDODS",
        "title": "Сервис аутентификации",
        "contact": {
            "name": "Evans Trein",
            "url": "https://github.com/EvansTrein",
            "email": "evanstrein@icloud.com"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "paths": {
...
        "/user/{guid}": {
            "get": {
                "description": "Получение данных пользователя с указанным GUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получение данных пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "format": "7c5e66cf-57ba-4871-9186-74ff5ab1e1f1",
                        "description": "Уникальный идентификатор пользователя",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserExample"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    }
                }
            }
        },
        "/user/{guid}/del": {
            "delete": {
                "description": "Удаление пользователя с указанным GUID и его сессии, если она есть",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Удаление пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "format": "7c5e66cf-57ba-4871-9186-74ff5ab1e1f1",
                        "description": "Уникальный идентификатор пользователя",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponceMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    }
                }
            }
        },
        "/user/{guid}/update": {
            "put": {
                "description": "Обновление данных пользователя по GUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Обновление данных пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "format": "7c5e66cf-57ba-4871-9186-74ff5ab1e1f1",
                        "description": "Уникальный идентификатор пользователя",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные пользователя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserDataUpd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponceMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrResponce": {
            "type": "object",
            "properties": {
                "errMessage": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "models.GetUserExample": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user1@mail.com"
                },
                "guid": {
                    "type": "string",
                    "example": "3c43e84d-fc44-4895-bc72-2a7924417b80"
                }
            }
        },
        "models.ResponceMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "info message"
                }
            }
        },
...
        "models.UserDataUpd": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user1@mail.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                }
            }
        }
    },
...
}